+++ ChangeLog.md
@@ -5,3 +5,4 @@
 - Initial version
 - Ajout du dictionnaire des spécifications de panneaux photovoltaïques
 - Préremplissage du dictionnaire PV avec les spécifications standards
+- Ajout des colonnes Type de caractéristiques et Position pour le dictionnaire PV